Add files via upload
THINGS ARE DONE!
Look for bugs and such to continue  polishing.
There is still one method that is too long and I'm afraid to touch it.
One thing I'm not a fan of is the blank square being red. It makes it hard to find at first glance.
Great job everybody!