Still needs changing
Need to break this into two separate class files. No idea how.
Look for bugs and such to continue  polishing.
Check_and_move is still too long and I'm afraid to touch it.
One thing I'm not a fan of is the blank square being red. It makes it hard to find at first glance.