95% complete
Changed tiles colors to a cooler shade.
Need to break this into two separate class files. No idea how.
Every function is 24 lines or less.
Look for bugs and such to continue  polishing.
One thing I'm not a fan of is the blank square being red. It makes it hard to find at first glance.